(1/2) Rename dyn_element.hh => dyn_element_draw.hh and pass slug to prepare_module
