+++ firmware/src/gui/pages/module_view.hh
@@ -1,7 +1,7 @@
 #pragma once
 #include "CoreModules/elements/element_info.hh"
 #include "conf/patch_conf.hh"
-#include "gui/dyn_element.hh"
+#include "gui/dyn_element_draw.hh"
 #include "gui/elements/element_name.hh"
 #include "gui/elements/map_ring_animate.hh"
 #include "gui/elements/module_drawer.hh"
@@ -296,7 +296,7 @@ struct ModuleViewPage : PageBase {
 			show_roller();
 		}
 
-		dyn_draw.prepare_module(this_module_id, canvas, 240);
+		dyn_draw.prepare_module(slug, this_module_id, canvas, 240);
 	}
 
 	void watch_element(DrawnElement const &drawn_element) {
@@ -430,8 +430,8 @@ struct ModuleViewPage : PageBase {
 				redraw_text_display(drawn_el, this_module_id, params.text_displays.watch_displays);
 			}
 
-			if (dyn_frame_throttle_ctr-- == 0) {
-				dyn_frame_throttle_ctr = DynFrameThrottle;
+			if (++dyn_frame_throttle_ctr >= DynFrameThrottle) {
+				dyn_frame_throttle_ctr = 0;
 				dyn_draw.draw();
 			}
 		}
@@ -556,8 +556,6 @@ private:
 	void reset_module_page() {
 		for (auto &b : button)
 			lv_obj_del(b);
-
-		dyn_draw.blur();
 
 		if (canvas)
 			lv_obj_del(canvas);
